Update Phrase_of_the_day.py
This script gets the phrase, definition and example of the "phrase of the day" from the website 'https://www.ihbristol.com/english-phrases/'. It then, saves it into an excel sheet called "PhrasesOfTheDay.xlsx" After that, it sends the former three things to a list of contacts named 'Contacts.txt'
The Contact.txt file has the next format:
Name1 Email1
Name2 Email2
It sends the mails using Yagmail.